Migrating Stacks' Cells to separate files
