+++ stack.c
@@ -2,35 +2,12 @@
 #include <stdio.h>
 #include "stack.h"
 
-void init_item(Item* item) {
-	// set default values for item
-	item->key = 0;
-	item->next = NULL;
-}
-
-Item* new_item() {
-	// allocate memory for a new item and initialize it
-	Item* item;
-	item = (Item*) malloc(sizeof(Item));
-	init_item(item);
-	return item;
-}
-
-void print_item(const Item* item) {
-	printf("%d\n", item->key);
-}
-
-void delete_item(Item* item) {
-	free(item);
-}
-
 void init_stack(Stack* stack) {
 	stack->top = NULL;
 	stack->size = 0;
 }
 
 Stack* new_stack() {
-	// allocate memory and initialize a stack
 	Stack *stack;
 	stack = (Stack*) malloc(sizeof(Stack));
 	init_stack(stack);
